delete fileloader trait and use new module system for loaders
