+++ src/loaders.rs
@@ -1,0 +1,2 @@
+pub mod gltf;
+pub mod obj;
