--- src/loaders/fileloader.rs
@@ -1,5 +1,0 @@
-use crate::geometry::Mesh;
-
-pub trait FileLoader {
-    fn load_file(filepath: &str) -> Vec<Mesh>;
-}
