--- src/loaders/mod.rs
@@ -1,8 +1,0 @@
-mod fileloader;
-pub use self::fileloader::*;
-
-mod gltfloader;
-pub use self::gltfloader::*;
-
-mod objloader;
-pub use self::objloader::*;
